InterviewIQ/requirements.txt
streamlit>=1.35.0
google-generativeai>=0.7.0
pdfplumber>=0.11.0
python-docx>=1.1.0
plotly>=5.22.0
pandas>=2.2.0
python-dotenv>=1.0.0